Changed min year args for AURN_processing to float from int.

+++ scripts/Data_Processing/AURN_processing.py
@@ -72,8 +72,8 @@ if __name__ == '__main__':
 
 
     parser.add_argument("--emep_filename","-e", default=None, help="filename of the emep file in CSV format (.csv)")
-    parser.add_argument("--min_years", "-n", type=int, help="minimum number of years of data that a site must have")
-    parser.add_argument("--min_years_ref", "-u", type=int, help="minimum number of years of data for any site that \
+    parser.add_argument("--min_years", "-n", type=float, help="minimum number of years of data that a site must have")
+    parser.add_argument("--min_years_ref", "-u", type=float, help="minimum number of years of data for any site that \
         we are going to use as a reference site later. (this cannot be less than min_years)")
     parser.add_argument("--sites", "-i", metavar='S', dest="sites", type=str, nargs='+', help="the measurement sites \
         to be processed. Default is to process all available AURN sites.")
